+++ src/main/java/net/mtrop/doom/bsp/BSPFormat.java
@@ -1,0 +1,31 @@
+package net.mtrop.doom.bsp;
+
+/**
+ * Enumeration of BSP format types.
+ * @author Matthew Tropiano
+ * @since [NOW]
+ */
+public enum BSPFormat 
+{
+	/** Regular Doom format. */
+	DOOM,
+	/** DeepBSPv4 format. */
+	DEEP, 
+	/** XNOD format. */
+	XNOD,
+	/** XGLN format. */
+	XGLN,
+	/** XGL2 format. */
+	XGL2,
+	/** XGL3 format. */
+	XGL3,
+	/** ZNOD format. */
+	ZNOD,
+	/** ZGLN format. */
+	ZGLN,
+	/** ZGL2 format. */
+	ZGL2,
+	/** ZGL3 format. */
+	ZGL3,
+	;
+}